feat: CalculateViewController 프로퍼티 정의
