--- Calculator/Calculator/CalculateViewController.swift
@@ -1,21 +1,0 @@
-//
-//  Calculator - CalculateViewController.swift
-//  Created by yagom. 
-//  Copyright © yagom. All rights reserved.
-// 
-
-import UIKit
-
-class CalculateViewController: UIViewController {
-
-    
-    @IBOutlet weak var enteringNumberLabel: UILabel!
-    @IBOutlet weak var enteringOperatorLabel: UILabel!
-    
-    
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
-
