+++ Calculator/Calculator/Model/CalculatorChecker.swift
@@ -1,0 +1,12 @@
+//
+//  CalculatorChecker.swift
+//  Calculator
+//
+//  Created by 천승현 on 2023/02/04.
+//
+
+import Foundation
+
+struct CalculatorChecker {
+    
+}
