+++ Calculator/Calculator/Model/Sign.swift
@@ -1,0 +1,12 @@
+//
+//  Sign.swift
+//  Calculator
+//
+//  Created by 천승현 on 2023/02/04.
+//
+
+enum Sign {
+    static let empty = ""
+    static let space = " "
+    static let zero = "0"
+}
